YG-112 🐞 #close #comment PR review 반영 완료
