+++ app/page.tsx
@@ -3,14 +3,14 @@
 import MainDesc from "./_components/mainDesc"
 import MainIntro from "./_components/mainIntro"
 import MainPosts from "./_components/mainPosts"
-import MainRecommend from "./_components/mainRecommend"
+import MainSurveyRecommend from "./_components/mainSurveyRecommend"
 import MainThemeBanner from "./_components/mainThemeBanner"
 
 const Home = () => {
     return (
         <main className=" overflow-x-hidden">
             <MainIntro />
-            <MainRecommend />
+            <MainSurveyRecommend />
             <MainDesc />
             <MainThemeBanner />
             <MainPosts />
