+++ components/layouts/_components/type.ts
@@ -1,0 +1,11 @@
+import { UserInfoProps } from "../type"
+
+export type HeaderLoginProps = {
+    isLoggedIn: boolean
+    userInfo: UserInfoProps | undefined
+}
+
+export type HeaderSearchBarProps = {
+    isSearchBarClicked: boolean
+    setIsSearchBarClicked: (isSearchBarClicked: boolean) => void
+}
